Make Ahoy.geocode apply everywhere
In 729d9e4496b3875 the ability to disable geocoding was introduced. It
only applied to those stores that callsed `set_visit_properties`, which
the LogStore, for example, does not.

This commit moves the removal of the LOCATION_KEYS from VisitProperties
into the VisitProperties class, so that it will apply everywhere.