allow ahoy.track_visit and ahoy.track_event to recieve blocks
This changes the Tracker so that the track_visit and track_event methods may receive blocks
and pass them to the store. Although the store methods of the same name would yield the
visit or event object, they were not passed a block from the Tracker instance.
This change also switches to block.call(object) instead of yield(object) because the
method signature already instructed the Ruby VM to create a block with the &block argument.